Create ETL Sparce Matrix Generator
Reduces the 2018 European Arrivals Dataset to a Sparse Matrix. The matrix has airport ICAO identifiers as the rows and columns. Each row represents an airport's busiest day, measured in arrival counts. The columns represent other airports flights may depart from en route to the target airport. A 1 in a row, col entry means a flight left from the col airport to the row airport on the row airport's busiest day.