Create Single_Airport_Biggest_Day_Search
Give the program an ICAO Code, and it will find the busiest day. It will then give the departure airports where flights left on route to the airport given. It calculates the average distance travelled to the target airport on its busiest day. It then visualizes this on a map of Europe.